+++ CHANGES.md
@@ -13,6 +13,14 @@
 
 ### misc
 
+## 2026.2.0
+
+**リリース日**: 2026-05-13
+
+- [ADD] shiguredo_vpl::list_adapters と AdapterSelector / AdapterInfo / PciAddress / MediaAdapterType を追加する
+  - @voluntas
+- [CHANGE] EncoderConfig::new / DecoderConfig::new / codec_info::supported_codecs にアダプタ指定を必須化する
+  - @voluntas
 
 ## 2026.1.2
 
@@ -22,7 +30,6 @@
   - 以前はアライメントされていないサイズでバッファを計算していたため、バッファサイズが足りずに SIGSEGV が発生していた
   - @melpon
 
-
 ## 2026.1.1
 
 **リリース日**: 2026-04-08
@@ -30,7 +37,6 @@
 - [FIX] パック済みフォーマットでも Y/U/V に値を設定する
   - mfxFrameData のドキュメントに、NV12 や YUY2 のようなフォーマットの場合であっても、Y/U/V をそれぞれ設定する必要があると書かれているため
   - @melpon
-
 
 ## 2026.1.0